Splitting Greeter and Notifier services apart
